Add get_ec2_id method
