+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "baram"
-version = "0.2.1"
+version = "0.2.2"
 description = "AWS Framework for python"
 authors = ["Kwangsik Lee <lks21c@gmail.com>"]
 readme = "readme.md"
